feat: split client pages for metadata
